+++ src/main/java/br/com/juliomakita/repository/IGenericDAO.java
@@ -1,7 +1,7 @@
 package br.com.juliomakita.repository;
 
 public interface IGenericDAO<T> {
-    public T get(Class<T> cl, Integer id);
+    public T get(Class<T> cl, Long id);
     public T save(T object);
     public void update(T object);
     public void delete(T object);
